+++ ChangeLog.md
@@ -6,6 +6,10 @@
 
 - Tweaked the styling of the command palette's scrollbar so it's less
   intrusive. ([#14](https://github.com/davep/textual-enhanced/pull/14))
+- Added `ChangeTheme` as a common command.
+  ([#15](https://github.com/davep/textual-enhanced/pull/15))
+- Added handling code for the common commands to `EnhancedScreen`.
+  ([#15](https://github.com/davep/textual-enhanced/pull/15))
 
 ## v0.4.0